requirements.txt
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dash==2.15.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==3.0.2
gunicorn==21.2.0
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
nest-asyncio==1.6.0
numpy==1.26.4
packaging==23.2
pandas==2.2.0
plotly==5.19.0
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0
retrying==1.3.4
six==1.16.0
tenacity==8.2.3
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.0
Werkzeug==3.0.1
zipp==3.17.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dash==2.15.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==3.0.2
gunicorn==21.2.0
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
nest-asyncio==1.6.0
numpy==1.26.4
packaging==23.2
pandas==2.2.0
plotly==5.19.0
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0
retrying==1.3.4
six==1.16.0
tenacity==8.2.3
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.0
Werkzeug==3.0.1
zipp==3.17.0
annotated-types==0.6.0
anyio==4.3.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dash==2.15.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
distro==1.9.0
exceptiongroup==1.2.0
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
nest-asyncio==1.6.0
numpy==1.26.4
openai==1.12.0
packaging==23.2
pandas==2.2.0
pillow==10.2.0
plotly==5.19.0
pydantic==2.6.2
pydantic_core==2.16.3
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
retrying==1.3.4
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.27
tenacity==8.2.3
tqdm==4.66.2
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.0
Werkzeug==3.0.1
zipp==3.17.0
annotated-types==0.6.0
anyio==4.3.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dash==2.15.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
distro==1.9.0
exceptiongroup==1.2.0
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
nest-asyncio==1.6.0
numpy==1.26.4
openai==1.12.0
packaging==23.2
pandas==2.2.0
pika==1.3.2
pillow==10.2.0
plotly==5.19.0
pydantic==2.6.2
pydantic_core==2.16.3
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
retrying==1.3.4
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.27
tenacity==8.2.3
tqdm==4.66.2
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.0
Werkzeug==3.0.1
zipp==3.17.0
annotated-types==0.6.0
anyio==4.3.0
blinker==1.7.0
boto3==1.34.50
botocore==1.34.50
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dash==2.15.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
distro==1.9.0
exceptiongroup==1.2.0
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
MarkupSafe==2.1.5
nest-asyncio==1.6.0
numpy==1.26.4
openai==1.12.0
packaging==23.2
pandas==2.2.0
pika==1.3.2
pillow==10.2.0
plotly==5.19.0
pydantic==2.6.2
pydantic_core==2.16.3
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
retrying==1.3.4
s3transfer==0.10.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.27
tenacity==8.2.3
tqdm==4.66.2
typing_extensions==4.9.0
tzdata==2024.1
urllib3==1.26.18
Werkzeug==3.0.1
zipp==3.17.0
